+++ lib/cloudinary/api.rb
@@ -1,30 +1,5 @@
-require 'rest_client'
-require 'json'
-
 class Cloudinary::Api
-  class Error < CloudinaryException; end
-  class NotFound < Error; end
-  class NotAllowed < Error; end
-  class AlreadyExists < Error; end
-  class RateLimited < Error; end
-  class BadRequest < Error; end
-  class GeneralError < Error; end
-  class AuthorizationRequired < Error; end
-
-  class Response < Hash
-    attr_reader :rate_limit_reset_at, :rate_limit_remaining, :rate_limit_allowed
-
-    def initialize(response=nil)
-      if response
-        # This sets the instantiated self as the response Hash
-        update Cloudinary::Api.parse_json_response response
-
-        @rate_limit_allowed   = response.headers[:x_featureratelimit_limit].to_i if response.headers[:x_featureratelimit_limit]
-        @rate_limit_reset_at  = Time.parse(response.headers[:x_featureratelimit_reset]) if response.headers[:x_featureratelimit_reset]
-        @rate_limit_remaining = response.headers[:x_featureratelimit_remaining].to_i if response.headers[:x_featureratelimit_remaining]
-      end
-    end
-  end
+  extend Cloudinary::Request
 
   def self.ping(options={})
     call_api(:get, "ping", {}, options)
@@ -348,220 +323,6 @@ class Cloudinary::Api
     call_json_api('GET', json_url, {}, 60, {})
   end
 
-  # Creates a new sub-account. Any users that have access to all sub-accounts will also automatically have access to the
-  # new sub-account.
-  # @param [String] name The display name as shown in the management console
-  # @param [String] cloud_name A case-insensitive cloud name comprised of alphanumeric and underscore characters.
-  #   Generates an error if the specified cloud name is not unique across all Cloudinary accounts.
-  #   Note: Once created, the name can only be changed for accounts with fewer than 1000 assets.
-  # @param [Object] custom_attributes Any custom attributes you want to associate with the sub-account
-  # @param [Boolean] enabled Whether to create the account as enabled (default is enabled)
-  # @param [String] base_account ID of sub-account from which to copy settings
-  # @param [Object] options additional options
-  def self.create_sub_account(name, cloud_name = nil, custom_attributes = {}, enabled = nil, base_account = nil, options = {})
-    params = {
-      name: name,
-      cloud_name: cloud_name,
-      custom_attributes: custom_attributes,
-      enabled: enabled,
-      base_sub_account_id: base_account
-    }
-
-    call_account_api(:post, 'sub_accounts', params, options.merge(content_type: :json))
-  end
-
-  # Updates the specified details of the sub-account.
-  # @param [String] sub_account_id The ID of the sub-account.
-  # @param [String] name The display name as shown in the management console
-  # @param [String] cloud_name A case-insensitive cloud name comprised of alphanumeric and underscore characters.
-  #   Generates an error if the specified cloud name is not unique across all Cloudinary accounts.
-  #   Note: Once created, the name can only be changed for accounts with fewer than 1000 assets.
-  # @param [Object] custom_attributes Any custom attributes you want to associate with the sub-account, as a map/hash
-  #   of key/value pairs.
-  # @param [Boolean] enabled Whether the sub-account is enabled.
-  # @param [Object] options additional options
-  def self.update_sub_account(sub_account_id, name = nil, cloud_name = nil, custom_attributes = nil, enabled = nil, options = {})
-    params = {
-      name: name,
-      cloud_name: cloud_name,
-      custom_attributes: custom_attributes,
-      enabled: enabled
-    }
-
-    call_account_api(:put, ['sub_accounts', sub_account_id], params, options.merge(content_type: :json))
-  end
-
-  # Lists sub-accounts.
-  # @param [Boolean] enabled Whether to only return enabled sub-accounts (true) or disabled accounts (false).
-  #  Default: all accounts are returned (both enabled and disabled).
-  # @param [Array<String>] ids A list of up to 100 sub-account IDs. When provided, other parameters are ignored.
-  # @param [String] prefix Returns accounts where the name begins with the specified case-insensitive string.
-  # @param [Object] options additional options
-  def self.sub_accounts(enabled = nil, ids = [], prefix = nil, options = {})
-    params = {
-      enabled: enabled,
-      ids: ids,
-      prefix: prefix
-    }
-
-    call_account_api(:get, 'sub_accounts', params, options.merge(content_type: :json))
-  end
-
-  # Retrieves the details of the specified sub-account.
-  # @param [String] sub_account_id The ID of the sub-account.
-  # @param [Object] options additional options
-  def self.sub_account(sub_account_id, options = {})
-    call_account_api(:get, ['sub_accounts', sub_account_id], {}, options.merge(content_type: :json))
-  end
-
-  # Deletes the specified sub-account. Supported only for accounts with fewer than 1000 assets.
-  # @param [String] sub_account_id The ID of the sub-account.
-  # @param [Object] options additional options
-  def self.delete_sub_account(sub_account_id, options = {})
-    call_account_api(:delete, ['sub_accounts', sub_account_id], {}, options)
-  end
-
-  # Creates a new user in the account.
-  # @param [String] name The name of the user.
-  # @param [String] email A unique email address, which serves as the login name and notification address.
-  # @param [String] role The role to assign. Possible values: master_admin, admin, billing, technical_admin, reports,
-  #   media_library_admin, media_library_user
-  # @param [Array<String>] sub_account_ids The list of sub-account IDs that this user can access.
-  #   Note: This parameter is ignored if the role is specified as master_admin.
-  # @param [Object] options additional options
-  def self.create_user(name, email, role, sub_account_ids = [], options = {})
-    params = {
-      name: name,
-      email: email,
-      role: role,
-      sub_account_ids: sub_account_ids
-    }
-
-    call_account_api(:post, 'users', params, options.merge(content_type: :json))
-  end
-
-  # Deletes an existing user.
-  # @param [String] user_id The ID of the user to delete.
-  # @param [Object] options additional options
-  def self.delete_user(user_id, options = {})
-    call_account_api(:delete, ['users', user_id], {}, options)
-  end
-
-  # Updates the details of the specified user.
-  # @param [String] user_id The ID of the user to update.
-  # @param [String] name The name of the user.
-  # @param [String] email A unique email address, which serves as the login name and notification address.
-  # @param [String] role The role to assign. Possible values: master_admin, admin, billing, technical_admin, reports,
-  #   media_library_admin, media_library_user
-  # @param [Array<String>] sub_account_ids The list of sub-account IDs that this user can access.
-  #   Note: This parameter is ignored if the role is specified as master_admin.
-  # @param [Object] options additional options
-  def self.update_user(user_id, name = nil, email = nil, role = nil, sub_account_ids = nil, options = {})
-    params = {
-      name: name,
-      email: email,
-      role: role,
-      sub_account_ids: sub_account_ids
-    }
-
-    call_account_api(:put, ['users', user_id], params, options.merge(content_type: :json))
-  end
-
-  # Returns the user with the specified ID.
-  # @param [String] user_id The ID of the user.
-  # @param [Object] options additional options
-  def self.user(user_id, options = {})
-    call_account_api(:get, ['users', user_id], {}, options.merge(content_type: :json))
-  end
-
-  # Lists users in the account.
-  # @param [Boolean] pending Whether to only return pending users. Default: all users
-  # @param [Array<String>] user_ids A list of up to 100 user IDs. When provided, other parameters are ignored.
-  # @param [String] prefix Returns users where the name or email address begins with the specified case-insensitive string.
-  # @param [String] sub_account_id Only returns users who have access to the specified account.
-  # @param [Object] options additional options
-  def self.users(pending = nil, user_ids = [], prefix = nil, sub_account_id = nil, options = {})
-    params = if user_ids && user_ids.count > 0
-               {
-                 ids: user_ids
-               }
-             else
-               {
-                 prefix: prefix,
-                 sub_account_id: sub_account_id,
-                 pending: pending
-               }
-             end
-
-    call_account_api(:get, 'users', params, options.merge(content_type: :json))
-  end
-
-  # Creates a new user group.
-  # @param [String] name The name for the user group.
-  # @param [Object] options additional options
-  def self.create_user_group(name, options = {})
-    params = {
-      name: name
-    }
-
-    call_account_api(:post, 'user_groups', params, options.merge(content_type: :json))
-  end
-
-  # Updates the specified user group.
-  # @param [String] group_id The ID of the user group to update.
-  # @param [String] name The name for the user group.
-  # @param [Object] options additional options
-  def self.update_user_group(group_id, name, options = {})
-    params = {
-      name: name
-    }
-
-    call_account_api(:put, ['user_groups', group_id], params, options.merge(content_type: :json))
-  end
-
-  # Adds a user to a group with the specified ID.
-  # @param [String] group_id The ID of the user group.
-  # @param [String] user_id The ID of the user.
-  # @param [Object] options additional options
-  def self.add_user_to_group(group_id, user_id, options = {})
-    call_account_api(:post, ['user_groups', group_id, 'users', user_id], {}, options.merge(content_type: :json))
-  end
-
-  # Removes a user from a group with the specified ID.
-  # @param [String] group_id The ID of the user group.
-  # @param [String] user_id The ID of the user.
-  # @param [Object] options additional options
-  def self.remove_user_from_group(group_id, user_id, options = {})
-    call_account_api(:delete, ['user_groups', group_id, 'users', user_id], {}, options.merge(content_type: :json))
-  end
-
-  # Deletes the user group with the specified ID.
-  # @param [String] group_id The ID of the user group to delete.
-  # @param [Object] options additional options
-  def self.delete_user_group(group_id, options = {})
-    call_account_api(:delete, ['user_groups', group_id], {}, options)
-  end
-
-  # Lists user groups in the account.
-  # @param [Object] options additional options
-  def self.user_groups(options = {})
-    call_account_api(:get, 'user_groups', {}, options.merge(content_type: :json))
-  end
-
-  # Retrieves the details of the specified user group.
-  # @param [String] group_id The ID of the user group to retrieve.
-  # @param [Object] options additional options
-  def self.user_group(group_id, options = {})
-    call_account_api(:get, ['user_groups', group_id], {}, options.merge(content_type: :json))
-  end
-
-  # Lists users in the specified user group.
-  # @param [String] group_id The ID of the user group.
-  # @param [Object] options additional options
-  def self.user_group_users(group_id, options = {})
-    call_account_api(:get, ['user_groups', group_id, 'users'], {}, options.merge(content_type: :json))
-  end
-
   protected
 
   def self.call_api(method, uri, params, options)
@@ -574,41 +335,6 @@ class Cloudinary::Api
     end
   end
 
-  def self.call_account_api(method, uri, params, options)
-    account_id = options[:account_id] || Cloudinary.config.account_id || raise('Must supply account_id')
-    api_key    = options[:provisioning_api_key] || Cloudinary.config.provisioning_api_key || raise('Must supply provisioning api_key')
-    api_secret = options[:provisioning_api_secret] || Cloudinary.config.provisioning_api_secret || raise('Must supply provisioning api_secret')
-
-    params.reject! { |_, v| v.nil? }
-
-    call_cloudinary_api(method, uri, api_key, api_secret, params, options) do |cloudinary, inner_uri|
-      [cloudinary, 'v1_1', 'provisioning', 'accounts', account_id, inner_uri]
-    end
-  end
-
-  def self.call_json_api(method, api_url, payload, timeout, headers, user = nil, password = nil)
-    RestClient::Request.execute(method: method,
-                                url: api_url,
-                                payload: payload,
-                                timeout: timeout,
-                                headers: headers,
-                                user: user,
-                                password: password) do |response|
-      return Response.new(response) if response.code == 200
-      exception_class = case response.code
-                        when 400 then BadRequest
-                        when 401 then AuthorizationRequired
-                        when 403 then NotAllowed
-                        when 404 then NotFound
-                        when 409 then AlreadyExists
-                        when 420 then RateLimited
-                        when 500 then GeneralError
-                        else raise GeneralError.new("Server returned unexpected status code - #{response.code} - #{response.body}")
-                        end
-      json = parse_json_response(response)
-      raise exception_class.new(json["error"]["message"])
-    end
-  end
   def self.parse_json_response(response)
     return Cloudinary::Utils.json_decode(response.body)
   rescue => e
@@ -659,23 +385,4 @@ class Cloudinary::Api
     params[by_key] = value
     call_api("post", "resources/#{resource_type}/#{type}/update_access_mode", params, options)
   end
-
-  def self.call_cloudinary_api(method, uri, user, password, params, options, &api_url_builder)
-    cloudinary = options[:upload_prefix] || Cloudinary.config.upload_prefix || 'https://api.cloudinary.com'
-    api_url    = Cloudinary::Utils.smart_escape(api_url_builder.call(cloudinary, uri).flatten.join('/'))
-    timeout    = options[:timeout] || Cloudinary.config.timeout || 60
-
-    headers = { 'User-Agent' => Cloudinary::USER_AGENT }
-
-    if options[:content_type] == :json
-      payload = params.to_json
-      headers.merge!("Content-Type" => "application/json", "Accept" => "application/json")
-    else
-      payload = params.reject { |_, v| v.nil? || v == "" }
-    end
-
-    call_json_api(method, api_url, payload, timeout, headers, user, password)
-  end
-
-  private_class_method :call_cloudinary_api
 end
